+++ spec/rails_app/db/migrate/20120219130149_add_published_to_post.rb
@@ -1,0 +1,5 @@
+class AddPublishedToPost < ActiveRecord::Migration
+  def change
+    add_column :posts, :published_at, :timestamp
+  end
+end